requirements.txt
annotated-types==0.7.0
anyio==4.5.2
click==8.1.8
et-xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
joblib==1.4.2
lxml==5.3.2
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
pydantic==2.10.6
pydantic-core==2.27.2
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
threadpoolctl==3.5.0
typing-extensions==4.13.2
tzdata==2025.2
uvicorn==0.33.0
